app/test: make parameters clearer when adding fast tests
For the fast tests, we have two extra parameters specifying when the
test can be run without hugepages or using ASan. The true/false nature
of these parameters is not very clear, so change things so that they are
explicitly specified as NOHUGE/HUGEPAGES and NOASAN/ASAN instead.
Explicitly validate the options in the meson.build files, rather than
just checking for one of the pair of options - which can hide errors.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>