net/iavf: use common Tx path selection infrastructure
Replace the existing complicated logic with the use of the common
function. Let the primary process select the Tx path to be used by all
processes using the given device.

Introduce two new features "disabled" and "context desc" to the common
infrastructure which represents whether or not the path is disabled, or
if it uses a context descriptor.

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>