net/idpf: use common Tx path selection infrastructure
Replace the existing complicated logic with the use of the common
function. Let the primary process select the Tx path to be used by all
processes using the given device.

Introduce a new feature "single queue" to the common infrastructure
which represents whether single or split queue mode is used in the given
path.

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>