net/intel: cache mbuf fast release mempool
When fast release of mbufs is enabled, the mempool to free the mbufs to
was determined by looking at the pool pointer of the first mbuf in the
burst being freed, potentially costing a cache miss.

This patch adds a mbuf fast release mempool pointer to the common transmit
queue structure, so reading the mbufs during fast release is avoided.
The pointer in located a cache line already being accessed,
and is only set once, when the first mbuf ever is released.

The fast release mempool pointer also indicates if fast release is
enabled, so this pointer is checked instead of the
RTE_ETH_TX_OFFLOAD_MBUF_FAST_FREE flag in the offloads field in the
transmit queue structure.

The same optimizations were applied to mbuf recycle.

For the ice driver, prefetcing the mbufs when fast release of mbufs is
enabled became superflouous, and has been moved into the branch for normal
mbuf release.

For the i40e driver, prefetcing the mbufs when fast release of mbufs is
enabled was already superflouous, and has been moved into the branch for
normal mbuf release.

Signed-off-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>