dts: make one link the default topology
Currently the default topology for test suites when not specified
is two link. However, the majority of the DPDK functions tested
in DTS can be tested with a single link. It is more intuitive for
a testsuite or testcase lacking a topology decorator to default
to the more basic requirement of 1 link, and force tests which
really do require two link to include decorators explicitly
saying so.

Signed-off-by: Dean Marx <dmarx@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>