net/sfc: avoid speed reset when setting FEC in started state
The 'current' mask of PHY capabilities in libefx indicates autonegotiation
regardless of whether the last invocation of 'efx_phy_adv_cap_set' enabled
or disabled it. The port start code is aware of this and pays attention to
preserve the existing setting when making 'efx_phy_adv_cap_set' invocation.

FEC set method uses the 'current' mask as a base for the new value and may
thus reset fixed speed to autonegotiated. Fix it to keep the speed setting.

Fixes: 5efa8fc1cfc0 ("net/sfc: support FEC feature")
Cc: stable@dpdk.org

Signed-off-by: Ivan Malov <ivan.malov@arknetworks.am>
Reviewed-by: Andy Moreton <andy.moreton@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>