net/gve: add reset path
Currently, the driver does not include any reset functionality and hence
cannot be reset by the application. This patch introduces a driver
callback for the `rte_eth_dev_reset` function which will reset the
device.

Also, as a precaution, null out device pointers to rx and tx queue
arrays when we release the queues during a reset.

Signed-off-by: Jasper Tran O'Leary <jtranoleary@google.com>
Reviewed-by: Joshua Washington <joshwash@google.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>